+++ functions/src/index.ts
@@ -13,3 +13,4 @@ export {saveGitHubToken} from "./handlers/serviceTokenHandlers";
 export {saveCard, getSavedCards} from "./handlers/savedCardHandlers";
 export {updateProfile} from "./handlers/profileHandlers";
 export {getPublicCard} from "./handlers/publicCardHandlers";
+export {manualSync} from "./handlers/syncHandlers";